Updates to home page content
Reading through this yesterday, I felt like the collaboration we get from Discord isn't highlighted enough. It seems to be a very useful component to making devBetter what it is. The collaboration, questions, discussions, etc there are great and happen just about on a daily basis.

This is to just spark some ideas of how to word this. Definitely open to additions to this.