Improve Video/Details Get performance by about 250ms
We are calling the vimeo endpoints `video/id` and `video/id/texttracks` simultaneously.
When they both return, we then call the url do get the texttrack we are interested in.
The `video/id` call is significantly longer than the `video/id/texttracks` one.
We only depend on the latter to continue with getting our texttrack.
Making that call as soon as possible saves us about 250ms.